Create app.js file that allows code to update in both android and ios simulataneously.
